add sae-sony-qrcode images
It is qrcode of my first front end works! It leads to the page I
designed and accomplished with HTML, CSS, JavaScript, jQuery - all kind
of FE skills after I analyzed the original 'Sony China' home page. It's
based on the original but developed with my own idea. Just for fun!